Long int

+++ CoSCode/CoSCode.ino
@@ -203,7 +203,7 @@ void loop() {
         break;    
       }
   }
-  int x = 1000000;
+  long int x = 1000000;
   step1.moveTo(x);
   step2.moveTo(-1 * x);
 }
